requirements.txt
atlassian-python-api~=3.25